% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputs.R
\name{check_est}
\alias{check_est}
\title{Check if enough estimates are given and if they are valid}
\usage{
check_est(x, param = c("RUD_p", "RUD_n", "CONS_p", "CONS_s", "CNT_s"))
}
\arguments{
\item{x}{a one row data.frame with columns "mean", "sd", "min", "max", "bg"}

\item{param}{character; parameter type}
}
\description{
Check if enough estimates are given and if they are valid
}
