% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defaults.R
\name{default_estimates}
\alias{default_estimates}
\title{Get default estimates for given parameters}
\usage{
default_estimates(
  species = c("Honey Bee", "Bumble Bee", "Solitary Bee"),
  stadium = c("adult", "larvae"),
  category = c("none", "forager", "nurse"),
  type = c("none", "chronic", "acute"),
  application = c("downwards", "up/sidewards")
)
}
\arguments{
\item{species}{character; species}

\item{stadium}{character; stadium}

\item{category}{character; category}

\item{type}{character; type}

\item{application}{character; application}
}
\description{
Get default estimates for given parameters
}
\examples{
default_estimates(species = "Honey Bee", stadium = "adult", category = "forager")
}
