% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyshval-package.R
\docType{data}
\name{defaults}
\alias{defaults}
\title{Default values for shinyshval}
\format{
A data frame with 24 rows and 10 variables:
\describe{
\item{species}{species}
\item{stadium}{stadium}
\item{category}{category}
\item{type}{type}
\item{application}{application}
\item{param}{param}
\item{mean}{mean}
\item{sd}{sd}
\item{min}{min}
\item{max}{max}
\item{bg}{bg}
}
}
\usage{
defaults
}
\description{
A dataset containing default values for consumptions & contents.
}
\examples{
\dontrun{
# save default from csv to /data
defaults <- read.table(
  system.file("extdata/default_estimates.csv", package = "shinyshval"),
  header = TRUE, sep = ';')
defaults$min <- as.numeric(defaults$min)
defaults$max <- as.numeric(defaults$max)
# usethis::use_data(defaults, overwrite = TRUE)
}
}
\keyword{datasets}
