% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputs.R
\name{est_to_momnts}
\alias{est_to_momnts}
\title{Convert estimates (mean + sd, min+max or best guess) to moments}
\usage{
est_to_momnts(
  x,
  param = c("RUD_p", "RUD_n", "CONS_p", "CONS_s", "CNT_s"),
  alpha = 0.01
)
}
\arguments{
\item{x}{a one row data.frame with columns "param", "mean", "sd", "min",
"max" & "bg"}

\item{param}{character; parameter type}

\item{alpha}{numeric; alpha parameter for estimation from range}
}
\value{
a data.frame with four columns:
a (=mean), b (=sd), dist (=ditribution name), and source (='raw')
}
\description{
The function converts various estimates to moments (
mean & sd, min & max or best guess) in this order.
The "param" (RUD_p, RUD_n, CONS_p, CONS_s, CNT_s) column is used to
determine the distribution.
For CNT_s a different order is used:
(min & max & bg, min & max or best guess)
}
