% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputs.R
\name{get_param}
\alias{get_param}
\title{Get and combine parameters from data.frames of raw data or estimates.}
\usage{
get_param(raw, est)
}
\arguments{
\item{raw}{data.frame with columns (RUD_p, RUD_n, CONS_p, CONS_s, CNT_s).}

\item{est}{data.frame with columns (mean, sd, min, max, bg) and
columns (RUD_p, RUD_n, CONS_p, CONS_s, CNT_s)}
}
\value{
a data.frame with four columns:
a (=mean), b (=sd), dist (=distribution name),
and source (='raw', or 'estimate')
and five rows (RUD_p, RUD_n, CONS_p, CONS_s, CNT_s)
}
\description{
Get and combine parameters from data.frames of raw data or estimates.
}
\examples{
raw <- structure(list(RUD_p = c(-0.87, -0.43, -0.26, 2.02, -0.27),
     RUD_n = c(1, 2, 3, 4, 5),
     CONS_p = c(NA_real_, NA_real_, NA_real_, NA_real_, NA_real_),
     CONS_s = c(NA_real_, NA_real_, NA_real_, NA_real_, NA_real_),
     CNT_s = c(NA_real_, NA_real_, NA_real_, NA_real_, NA_real_)),
   .Names = c("RUD_p", "RUD_n", "CONS_p", "CONS_s",  "CNT_s"),
   row.names = c(NA, 5L),
   class = "data.frame")
est <- structure(list(
     param = c("RUD_p", "RUD_n", "CONS_p", "CONS_s", "CNT_s"),
     mean = c(NA_real_, NA_real_, NA_real_, NA_real_, NA_real_),
     sd = c(NA_real_, NA_real_, NA_real_, NA_real_, NA_real_),
     min = c(NA, NA, 1, NA, 0.1),
     max = c(NA, NA, 3, NA, 0.5),
     bg = c(NA, NA, NA, 5, 0.4)),
 .Names = c("param", "mean", "sd", "min", "max", "bg"),
 row.names = c(NA, -5L),
 class = "data.frame")
 get_param(raw, est)
}
