% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{parse_raw}
\alias{parse_raw}
\title{helper function to parse original input files}
\usage{
parse_raw(file)
}
\arguments{
\item{file}{character; filepath}
}
\description{
helper function to parse original input files
}
\examples{
parse_raw(system.file("extdata/original.csv", package = "shinyshval"))
}
