% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputs.R
\name{raw_to_momnts}
\alias{raw_to_momnts}
\title{Convert raw measurements to moments}
\usage{
raw_to_momnts(x, param = c("RUD_p", "RUD_n", "CONS_p", "CONS_s", "CNT_s"))
}
\arguments{
\item{x}{numeric; raw data}

\item{param}{character; parameter type}
}
\value{
a data.frame with four columns:
a (=mean), b (=sd), dist (=ditribution name), and source (='raw')
}
\description{
a normal distribution is fitted - except for CNT_s,
where a beta distribution is fitted.
}
\examples{
raw <- c(-0.87, -0.43, -0.26, 2.02, -0.27)
raw_to_momnts(raw, "RUD_p")
raw2 <- c(0.94, 0.91, 0.79, 0.62, 0.83)
raw_to_momnts(raw2, "CNT_s")
}
