% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.R
\name{sim}
\alias{sim}
\title{Compute 90th percentile of shval}
\usage{
sim(
  param,
  nsim = 1000,
  iter = nsim,
  q = 0.9,
  twa = FALSE,
  dt50_p = 2,
  dt50_n = 2,
  t = 10
)
}
\arguments{
\item{param}{data.frame; a data.frame holding th estimates,
as returned by \code{\link[=get_param]{get_param()}}}

\item{nsim}{integer; number of simulation to perform}

\item{iter}{integer; number of iterations}

\item{q}{numeric; with quantile should be returned?}

\item{twa}{logical; should a time-weighted-average (twa) be returned?}

\item{dt50_p}{numeric; needed for twa}

\item{dt50_n}{numeric; needed for twa}

\item{t}{numeric; needed for twa}
}
\value{
a numeric vector
}
\description{
Compute 90th percentile of shval
}
