% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defaults.R
\name{default_estimates}
\alias{default_estimates}
\title{Get default estimates for given parameters}
\usage{
default_estimates(
  species = c("Honey Bee", "Bumble Bee", "Solitary Bee"),
  stadium = c("adult", "larvae"),
  category = c("none", "forager", "nurse"),
  type = c("none", "chronic", "acute"),
  application = c("downwards", "up/sidewards")
)
}
\arguments{
\item{species}{character; species}

\item{stadium}{character; stadium}

\item{category}{character; category}

\item{type}{character; type}

\item{application}{character; application}
}
\description{
Defaults where extracted from tables J4-J8 of the guideline, as good as possible (see note).
}
\note{
RUD values are given as mean +- sd on raw scale. However, for shval these numbers on
log-scale are needed. For the upwards scenarious we used the mean and sd from the original data provided,
which fits to the guideline.
For the sidewards scenarious the sd was not possible to exract and the sd from the raw-scale was used!
This gives a slighly higher sd.
}
\examples{
default_estimates(species = "Honey Bee",
  stadium = "adult",
  category = "forager")
}
